[Refactoring] Rename database to manager.
