+++ app/src/main/java/engineer/kaobei/fragment/DashBoardFragment.kt
@@ -27,7 +27,7 @@ import engineer.kaobei.activity.ArticleActivity
 import engineer.kaobei.activity.LoginActivity
 import engineer.kaobei.activity.MainActivity
 import engineer.kaobei.activity.SettingsActivity
-import engineer.kaobei.database.AuthStateManager
+import engineer.kaobei.manager.AuthStateManager
 import engineer.kaobei.model.articles.Article
 import engineer.kaobei.model.kaobeluser.KaobeiUser
 import engineer.kaobei.model.userarticles.UserArticle
