+++ app/src/main/java/engineer/kaobei/manager/OpenSourceManager.kt
@@ -1,4 +1,4 @@
-package engineer.kaobei.database
+package engineer.kaobei.manager
 
 import android.content.Context
 import com.google.gson.Gson
